rag_service/requirements.txt
fastapi
uvicorn
python-multipart
requests
openai
chromadb
sentence-transformers
python-dotenv
